test(galgame): optimize bridge test setup
