+++ plugin/tests/unit/plugins/_galgame_agent_support.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+
+from _galgame_bridge_support import (
+    _BlockingSummaryGateway,
+    _drain_agent_summary_tasks,
+    _FakeHostAdapter,
+    _FakeLLMGateway,
+    _run_in_new_loop,
+    _summary_test_line,
+    _summary_test_line_event,
+)
+
+__all__ = [name for name in globals() if not name.startswith("__")]
