+++ pytest.ini
@@ -2,7 +2,7 @@
 asyncio_mode = auto
 asyncio_default_fixture_loop_scope = function
 addopts = -p no:anyio
-norecursedirs = local_server .git __pycache__ *.egg
+norecursedirs = local_server .git __pycache__ *.egg .codex-* .codex-tmp .codex_pytest_tmp .codex_manual_checks .pytest-* pytest-cache-files-* tmp-pytest N.E.K.O
 markers =
     unit: mark a test as a unit test.
     frontend: mark a test as a frontend test.
